fix: neatness refactor
