+++ packages/browser/lib/curves/p256.ts
@@ -1,4 +1,6 @@
-import { createCurve } from "./base";
+import createCurve from "./createCurve";
+
+export type TP256Constants = typeof P256_CONSTANTS;
 
 const P256_CONSTANTS = {
   p: "FF FF FF FF 00 00 00 01 00 00 00 00 00 00 00 00 00 00 00 00 FF FF FF FF FF FF FF FF FF FF FF FF",
@@ -10,5 +12,6 @@ const P256_CONSTANTS = {
   n: "FF FF FF FF 00 00 00 00 FF FF FF FF FF FF FF FF BC E6 FA AD A7 17 9E 84 F3 B9 CA C2 FC 63 25 51",
   h: "01",
 };
+
 
 export const encodePublicKey = createCurve(P256_CONSTANTS);
